Split product formats into refelective and non-reflective parts
