+++ shared/src/main/boilerplate/spray/json/ReflectiveProductFormatsInstances.scala.template
@@ -16,26 +16,12 @@
 
 package spray.json
 
-trait ProductFormatsInstances { self: ProductFormats with StandardFormats =>
+trait ReflectiveProductFormatsInstances { self: ReflectiveProductFormats with StandardFormats =>
 [#  // Case classes with 1 parameters
 
   def jsonFormat1[[#P1 :JF#], T <: Product :ClassManifest](construct: ([#P1#]) => T): RootJsonFormat[T] = {
     val Array([#p1#]) = extractFieldNames(classManifest[T])
     jsonFormat(construct, [#p1#])
-  }
-  def jsonFormat[[#P1 :JF#], T <: Product](construct: ([#P1#]) => T, [#fieldName1: String#]): RootJsonFormat[T] = new RootJsonFormat[T]{
-    def write(p: T) = {
-      val fields = new collection.mutable.ListBuffer[(String, JsValue)]
-      fields.sizeHint(1 * 2)
-      [#fields ++= productElement##2Field[P1](fieldName1, p, 0)#
-      ]
-      JsObject(fields: _*)
-    }
-    def read(value: JsValue) = {
-      [#val p1V = fromField[P1](value, fieldName1)#
-      ]
-      construct([#p1V#])
-    }
   }#
