+++ shared/src/main/scala/spray/json/CompactPrinter.scala
