+++ shared/src/main/scala/spray/json/ReflectiveProductFormats.scala
@@ -1,68 +1,14 @@
-/*
- * Copyright (C) 2011 Mathias Doenitz
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package spray.json
 
 import java.lang.reflect.Modifier
+
 import scala.annotation.tailrec
+import scala.reflect.ClassManifest
 import scala.util.control.NonFatal
 
-/**
- * Provides the helpers for constructing custom JsonFormat implementations for types implementing the Product trait
- * (especially case classes)
- */
-trait ProductFormats extends ProductFormatsInstances {
+
+trait ReflectiveProductFormats extends ProductFormats with ReflectiveProductFormatsInstances {
   this: StandardFormats =>
-
-  def jsonFormat0[T](construct: () => T): RootJsonFormat[T] =
-    new RootJsonFormat[T] {
-      def write(p: T) = JsObject()
-      def read(value: JsValue) = value match {
-        case JsObject(_) => construct()
-        case _ => throw new DeserializationException("Object expected")
-      }
-    }
-
-  // helpers
-  
-  protected def productElement2Field[T](fieldName: String, p: Product, ix: Int, rest: List[JsField] = Nil)
-                                       (implicit writer: JsonWriter[T]): List[JsField] = {
-    val value = p.productElement(ix).asInstanceOf[T]
-    writer match {
-      case _: OptionFormat[_] if (value == None) => rest
-      case _ => (fieldName, writer.write(value)) :: rest
-    }
-  }
-
-  protected def fromField[T](value: JsValue, fieldName: String)
-                                     (implicit reader: JsonReader[T]) = value match {
-    case x: JsObject if
-      (reader.isInstanceOf[OptionFormat[_]] &
-        !x.fields.contains(fieldName)) =>
-      None.asInstanceOf[T]
-    case x: JsObject =>
-      try reader.read(x.fields(fieldName))
-      catch {
-        case e: NoSuchElementException =>
-          deserializationError("Object is missing required member '" + fieldName + "'", e, fieldName :: Nil)
-        case DeserializationException(msg, cause, fieldNames) =>
-          deserializationError(msg, cause, fieldName :: fieldNames)
-      }
-    case _ => deserializationError("Object expected in field '" + fieldName + "'", fieldNames = fieldName :: Nil)
-  }
 
   protected def extractFieldNames(classManifest: ClassManifest[_]): Array[String] = {
     val clazz = classManifest.erasure
@@ -79,7 +25,7 @@ trait ProductFormats extends ProductFormatsInstances {
         sys.error("Case class " + clazz.getName + " declares additional fields")
       if (fields.zip(copyDefaultMethods).exists { case (f, m) => f.getType != m.getReturnType })
         sys.error("Cannot determine field order of case class " + clazz.getName)
-      fields.map(f => ProductFormats.unmangle(f.getName))
+      fields.map(f => ReflectiveProductFormats.unmangle(f.getName))
     } catch {
       case NonFatal(ex) => throw new RuntimeException("Cannot automatically determine case class field names and order " +
         "for '" + clazz.getName + "', please use the 'jsonFormat' overload with explicit field name specification", ex)
@@ -88,7 +34,7 @@ trait ProductFormats extends ProductFormatsInstances {
 
 }
 
-object ProductFormats {
+object ReflectiveProductFormats {
 
   private def unmangle(name: String) = {
     import java.lang.{StringBuilder => JStringBuilder}
@@ -133,22 +79,5 @@ object ProductFormats {
       } else if (builder != null) builder.toString else name
     }
     rec(0, null)
-  }
-}
-
-/**
- * This trait supplies an alternative rendering mode for optional case class members.
- * Normally optional members that are undefined (`None`) are not rendered at all.
- * By mixing in this trait into your custom JsonProtocol you can enforce the rendering of undefined members as `null`.
- * (Note that this only affect JSON writing, spray-json will always read missing optional members as well as `null`
- * optional members as `None`.)
- */
-trait NullOptions extends ProductFormats {
-  this: StandardFormats =>
-
-  override protected def productElement2Field[T](fieldName: String, p: Product, ix: Int, rest: List[JsField])
-                                                (implicit writer: JsonWriter[T]) = {
-    val value = p.productElement(ix).asInstanceOf[T]
-    (fieldName, writer.write(value)) :: rest
   }
 }
