+++ shared/src/test/scala/spray/json/CollectionFormatsSpec.scala
