+++ shared/src/test/scala/spray/json/CompactPrinterSpec.scala
