+++ shared/src/test/scala/spray/json/CustomFormatSpec.scala
